+++ lib/deprecool.rb
@@ -6,6 +6,10 @@ module Deprecool
   class Error < StandardError; end
 end
 
+Dir["#{__dir__}/deprecool/helpers/*.rb"].each do |helper|
+  require helper
+end
+
 require_relative 'deprecool/version'
 require_relative 'deprecool/finder'
 require_relative 'deprecool/registry'
@@ -15,7 +19,6 @@ require_relative 'deprecool/cli'
 
 # TODO: Scan gemfiles and then only require relevant finders instead of all
 # of them
-Dir[File.join(__dir__, 'deprecool', 'finders', '**', '*.rb')].each do |finder|
+Dir["#{__dir__}/deprecool/finders/**/*.rb"].each do |finder|
   require finder
 end
-