+++ lib/deprecool/finders/rails/v8.2.0/no_binds_to_insert_update_delete.rb
@@ -23,7 +23,7 @@ module Deprecool
             return unless arguments.length > 1
 
             confidence = confidence_from_arguments(arguments)
-            return if confidence == :none
+
 
             add_offense(node, confidence: confidence)
           end